… test/rpc-integration/resources/simplify.kore
… test/rpc-integration/test-simplify/response-if-then-else-arity-error.json
… test/rpc-integration/test-simplify/response-if-then-else-eval.json
… test/rpc-integration/test-simplify/response-if-then-else-false.json
… test/rpc-integration/test-simplify/response-if-then-else-indeterminate.json
… test/rpc-integration/test-simplify/response-if-then-else-sort-error.json
… test/rpc-integration/test-simplify/response-if-then-else-true.json
… test/rpc-integration/test-simplify/response-with-logging.json
… test/rpc-integration/test-simplify/state-if-then-else-indeterminate.simplify
… test/rpc-integration/test-simplify/state-if-then-else-sort-error.simplify
… test/rpc-integration/test-simplify/state-if-then-else-true.simplify