* Evaluation now proceeds bottom-up (but still discovers work top-down)
* New module `Booster.Builtin` as a home for all hook implementations
* If a hook implementation is provided, it will be tried before function and simplification equations
  * A hook is allowed to return no result, in which case function and simplification equations will be tried
  * Sort errors or arity errors (currently) abort the evaluation (should be caught during internalisation)

FIxes #437