@jberthold 2023-12-21
KEVM tests show no change - even an empty file when comparing with default

| Test                                     | 437-if-then-else-hook time | master-4a693bc time | (437-if-then-else-hook/master-4a693bc) time 
|------------------------------------------|----------------------------|---------------------|---------------------------------------------
| erc20/ds/transferFrom-success-2-spec.k   | 95.83                      | 98.5                | 0.9728934010152284                          
| erc20/ds/transfer-failure-2-a-spec.k     | 69.37                      | 70.64               | 0.982021517553794                           
| benchmarks/storagevar02-overflow-spec.k  | 88.71                      | 90.19               | 0.9835901984698968                          
| mcd/vat-muluu-pass-spec.k                | 107.82                     | 109.58              | 0.9839386749406825                          
| bihu/forwardToHotWallet-failure-2-spec.k | 132.66                     | 134.74              | 0.9845628618079263                          
| erc20/hkg/transfer-success-2-spec.k      | 51.47                      | 52.27               | 0.9846948536445379                          
| benchmarks/encode-keccak00-spec.k        | 53.31                      | 54.06               | 0.9861265260821309                          
| bihu/functional-spec.k                   | 116.47                     | 118.0               | 0.9870338983050847                          
| benchmarks/requires01-a0gt0-spec.k       | 51.55                      | 52.19               | 0.9877371143897298                          
| benchmarks/storagevar00-spec.k           | 64.63                      | 65.42               | 0.987924182207276                           
| erc20/ds/transferFrom-failure-1-c-spec.k | 123.77                     | 125.15              | 0.9889732321214542                          
| benchmarks/staticloop00-a0lt10-spec.k    | 53.0                       | 53.57               | 0.9893597162591002                          
| benchmarks/encodepacked-keccak01-spec.k  | 48.82                      | 49.32               | 0.9898621248986212                          
| erc20/ds/transferFrom-failure-1-b-spec.k | 153.09                     | 154.63              | 0.9900407424173835                          
| erc20/hkg/transferFrom-failure-1-spec.k  | 104.87                     | 103.83              | 1.0100163729172686                          
| erc20/ds/transfer-success-1-spec.k       | 73.83                      | 73.08               | 1.0102627257799672                          
| mcd/flipper-bids-pass-rough-spec.k       | 107.79                     | 106.63              | 1.0108787395667262                          
| bihu/forwardToHotWallet-failure-3-spec.k | 102.38                     | 101.15              | 1.0121601581809194                          
| erc20/ds/transferFrom-failure-1-d-spec.k | 103.59                     | 101.97              | 1.0158870255957635                          
| TOTAL                                    | 1702.9599999999998         | 1714.9199999999998  | 0.9930259137452476  

kontrol tests got borken by latest merges making booster default (the `--use-booster` is rejected), rerunning now