@jberthold 2023-12-21
| Test                                              | 437-if-then-else-hook time | master-4a693bc time | (437-if-then-else-hook/master-4a693bc) time 
|---------------------------------------------------|----------------------------|---------------------|---------------------------------------------
| BlockParamsTest.testBlockNumber()                 | 31.43                      | 33.67               | 0.9334719334719335                          
| AssumeTest.test_assume_true(uint256,uint256)      | 29.82                      | 31.24               | 0.9545454545454546                          
| AccountParamsTest.test_Nonce_NonExistentAddress() | 43.57                      | 45.53               | 0.9569514605754448                          
| AddrTest.test_notBuiltinAddress_symbolic(address) | 46.2                       | 48.16               | 0.9593023255813955                          
| ExpectRevertTest.testFail_expectRevert_empty()    | 30.68                      | 31.95               | 0.9602503912363067                          
| GasTest.testSetGas()                              | 25.69                      | 24.48               | 1.0494281045751634                          
| TOTAL                                             | 207.39                     | 215.02999999999997  | 0.9644700739431707