@jberthold 2023-12-21
Open question in this context: (Why) Do we need to have the `K` sort on the rules for `#if` in `domains.md`? 
https://github.com/runtimeverification/k/blob/master/k-distribution/include/kframework/builtin/domains.md?plain=1#L2304-L2305