@ehildenb 2023-12-21
I'm pretty sure the old backend actually applies hooks _after_ user-supplied simplifications. This is because the use may want to do something more structured than what the hook does, by recognizing a specific pattern and rewriting it to something else.

It may be good to follow the same behavior, but good to check first.