@ehildenb 2023-12-21
Bascially, if the user supplies rules for a particular scenario, we should use those rules. This is under the general principle of "the tool should be responsive to user input, rather than relying on magic".