@jberthold 2024-01-08
> I'm pretty sure the old backend actually applies hooks _after_ user-supplied simplifications. This is because the use may want to do something more structured than what the hook does, by recognizing a specific pattern and rewriting it to something else.

At the moment, the old backend [applies hooks before user-supplied equations (simplifications and equations)](https://github.com/runtimeverification/haskell-backend/blob/master/kore/src/Kore/Rewrite/Function/Evaluator.hs#L316-L321). 

That said, the idea to follow the user's guidelines first before using a hook is definitely worth considering. We just need to make sure there are no hidden issues from rules in the K prelude (`domains.md` ) not being coherent to the hook behaviour.